+++ packages/gasket-resolve/test/package-identifier.test.js
@@ -256,11 +256,9 @@ describe('projectIdentifier', () => {
     expect(result).toContain(type);
   });
 
-  it('generates unique names for anonymous entities', () => {
+  it('throws if a name is not provided', () => {
     const identifier = projectIdentifier('gasket', 'plugin');
-    const name1 = identifier().fullName;
-    const name2 = identifier().fullName;
-    expect(name1).not.toEqual(name2);
+    expect(identifier).toThrow(Error);
   });
 
   describe('isValidFullName', () => {